[Express Checkout] Split JS into shared core and cart bootstrap
